chore: refactor for easier reasoning
