+++ plugins/plugin-chart-handlebars/src/components/ControlHeader/controlHeader.tsx
@@ -1,0 +1,17 @@
+import React, { ReactNode } from 'react';
+
+interface ControlHeaderProps {
+  children: ReactNode;
+}
+
+export const ControlHeader = ({ children }: ControlHeaderProps): JSX.Element => {
+  return (
+    <div className="ControlHeader">
+      <div className="pull-left">
+        <label>
+          <span role={'button'}>{children}</span>
+        </label>
+      </div>
+    </div>
+  );
+};
