+++ plugins/plugin-chart-handlebars/src/plugin/controls/limits.ts
@@ -1,0 +1,16 @@
+import { ControlPanelsContainerProps, ControlSetItem } from '@superset-ui/chart-controls';
+import { isAggMode } from './shared';
+
+export const RowLimitControlSetItem: ControlSetItem = {
+  name: 'row_limit',
+  override: {
+    visibility: ({ controls }: ControlPanelsContainerProps) => !controls?.server_pagination?.value,
+  },
+};
+
+export const TimeSeriesLimitMetricControlSetItem: ControlSetItem = {
+  name: 'timeseries_limit_metric',
+  override: {
+    visibility: isAggMode,
+  },
+};
